refactor : hook and component divide
