+++ src/features/post/model/hooks/useDeletePostQuery.ts
@@ -1,0 +1,13 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query"
+import { deletePost } from "../../../../shared/api/posts"
+
+export const useDeletePostQuery = () => {
+  const queryClient = useQueryClient()
+
+  useMutation({
+    mutationFn: deletePost,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["posts"] })
+    },
+  })
+}
